Feat: /admin at Admin middleware

+++ routes/api.php
@@ -82,27 +82,25 @@ Route::middleware(['auth:investor'])->group(function () {
 
 // Rotas administrativas protegidas por auth:api e verificação de administrador
 Route::middleware(['auth:api','isAdmin'])->group(function() {
-    Route::get('investments', [InvestmentController::class, 'index']);
-    Route::get('investors', [InvestorController::class, 'index']);
-    Route::get('investors/{id}', [InvestorController::class, 'show']);
-    Route::put('investors/{id}', [InvestorController::class, 'update']);
-    Route::delete('investors/{id}', [InvestorController::class, 'destroy']);
+    Route::get('admin/investments', [InvestmentController::class, 'index']);
+    Route::get('admin/investors', [InvestorController::class, 'index']);
+    Route::get('admin/investors/{id}', [InvestorController::class, 'show']);
+    Route::put('admin/investors/{id}', [InvestorController::class, 'update']);
+    Route::delete('admin/investors/{id}', [InvestorController::class, 'destroy']);
     Route::post('admin/properties', [PropertyController::class, 'store']);
-    Route::get('properties/{id}/tokens', [PropertyController::class, 'tokens']);
-    Route::post('properties/{id}/tokenize', [PropertyController::class, 'tokenize']);
+    Route::get('admin/properties/{id}/tokens', [PropertyController::class, 'tokens']);
+    Route::post('admin/properties/{id}/tokenize', [PropertyController::class, 'tokenize']);
     Route::get('admin/properties', [PropertyController::class, 'index']);
 
-    Route::get('user/profile', [UserController::class, 'profile']);
+    Route::get('admin/user/profile', [UserController::class, 'profile']);
     Route::get('admin/imoveis/{id}/financeiro', [PropertyFinanceController::class, 'report']);
     Route::post('admin/imoveis/{id}/buyback', [BuybackController::class, 'buyback']);
 
     // Configurações da plataforma
    
-    Route::put('platform-settings', [PlatformSettingsController::class, 'update']);
+    Route::put('admin/platform-settings', [PlatformSettingsController::class, 'update']);
     Route::get('admin/transacoes-financeiras', [TransacaoFinanceiraController::class, 'lista']);
     // Route::get('admin/properties', [PropertyController::class, 'index']);
-
-
 });
 
 // Funcionalidades disponíveis para investidores autenticados
